Remove generation company/repo fields; add site URL and shared settings form
