+++ apps/web/src/components/directories/detail/generator/GeneratorForm.tsx
@@ -10,7 +10,6 @@ import {
 } from '@/lib/api/types-only';
 import { RequiredFields } from './RequiredFields';
 import { UpdateItemsFields } from './UpdateItemsFields';
-import { CompanyFields } from './CompanyFields';
 import { DynamicPluginFields } from './DynamicPluginFields';
 import { Button } from '@/components/ui/button';
 import { toast } from 'sonner';
@@ -26,7 +25,6 @@ import { generateItems, updateItems } from '@/app/actions/dashboard/generator';
 import { useTranslations } from 'next-intl';
 import { GenerationMethod, WebsiteRepositoryCreationMethod } from '@/lib/api/enums';
 import { getFormSchema } from '@/app/actions/dashboard/generator-form';
-import { CollapsibleSection } from '../shared';
 import { useProviderSelection } from '@/lib/hooks/use-provider-selection';
 import { ProviderSelectionSection } from '@/components/directories/shared/ProviderSelectionSection';
 
@@ -56,16 +54,12 @@ export function GeneratorForm({ directoryId, directory, config }: GeneratorFormP
     const [coreData, setCoreData] = useState<{
         name: string;
         prompt: string;
-        company?: { name: string; website: string };
-        repository_description?: string;
         generation_method?: GenerationMethod;
         update_with_pull_request?: boolean;
         website_repository_creation_method?: WebsiteRepositoryCreationMethod;
     }>({
         name: directory.name,
         prompt: initialPrompt,
-        company: lastRequestData?.company || undefined,
-        repository_description: lastRequestData?.repository_description || '',
         generation_method: GenerationMethod.CREATE_UPDATE,
         update_with_pull_request: false,
         website_repository_creation_method:
@@ -193,8 +187,6 @@ export function GeneratorForm({ directoryId, directory, config }: GeneratorFormP
                 const generateData: CreateItemsGeneratorDto = {
                     name: coreData.name,
                     prompt: coreData.prompt,
-                    company: coreData.company,
-                    repository_description: coreData.repository_description,
                     generation_method: coreData.generation_method,
                     update_with_pull_request: coreData.update_with_pull_request,
                     website_repository_creation_method: coreData.website_repository_creation_method,
@@ -262,18 +254,6 @@ export function GeneratorForm({ directoryId, directory, config }: GeneratorFormP
                             onProviderChange={handleProviderChange}
                         />
                     )}
-
-                    {/* Company Information */}
-                    <CollapsibleSection
-                        title={t('companyInformation')}
-                        description={t('companyInfoDescription')}
-                        defaultExpanded={false}
-                    >
-                        <CompanyFields
-                            company={coreData.company}
-                            onChange={(company) => handleCoreDataChange({ company })}
-                        />
-                    </CollapsibleSection>
 
                     {/* Dynamic Plugin Fields */}
