+++ apps/web/src/lib/api/directory.ts
@@ -279,6 +279,7 @@ export interface WebsiteSettings {
 
 export interface WebsiteSettingsResponse {
     company_name: string;
+    company_website: string;
     settings: WebsiteSettings;
     custom_menu: {
         header: CustomMenuItem[];
@@ -288,6 +289,7 @@ export interface WebsiteSettingsResponse {
 
 export interface UpdateWebsiteSettingsDto {
     company_name?: string;
+    company_website?: string;
     categories_enabled?: boolean;
     companies_enabled?: boolean;
     tags_enabled?: boolean;
