+++ packages/agent/package.json
@@ -161,6 +161,7 @@
         "class-validator": "^0.14.3",
         "date-fns": "^4.1.0",
         "deepmerge": "^4.3.1",
+        "github-slugger": "^2.0.0",
         "import-in-the-middle": "^2.0.1",
         "isomorphic-git": "^1.36.1",
         "lodash": "^4.17.21",
