+++ packages/agent/src/services/directory-generation.service.ts
@@ -1030,7 +1030,6 @@ export class DirectoryGenerationService {
 
             if ((generated.stats?.totalItemsCount ?? 0) > 0) {
                 await this.markdownGenerator.initialize(directory, user, {
-                    repository_description: dto.repository_description,
                     generation_method: dto.generation_method,
                     pr_update: generated.prUpdate,
                 });
