--- packages/contracts/src/api/generator/company.dto.ts
@@ -1,9 +1,0 @@
-/**
- * Company information for item generation
- */
-export interface CompanyDto {
-	/** Company name */
-	name: string;
-	/** Company website URL */
-	website: string;
-}
