+++ packages/contracts/src/api/generator/index.ts
@@ -2,7 +2,6 @@
 export { GenerationMethod, WebsiteRepositoryCreationMethod } from './generation-method.enum.js';
 
 // Generator DTOs
-export type { CompanyDto } from './company.dto.js';
 export type { ProvidersDto } from './providers.dto.js';
 export type { CreateItemsGeneratorDto, UpdateItemsGeneratorDto } from './create-items-generator.dto.js';
 export type { SubmitItemDto } from './submit-item.dto.js';
